Lambda parser
GitOrigin-RevId: d7dfa5b2873e95417e60e973be901264ce7adc69

+++ gdscript/src/main/kotlin/gdscript/parser/expr/GdFuncDeclExParser.kt
@@ -28,9 +28,10 @@ object GdFuncDeclExParser : GdExprBaseParser {
         ok = ok && b.consumeToken(COLON, true)
         ok = ok && GdStmtParser.parseLambda(b, false, true)
 
-        m.drop()
+        if (ok) m.drop()
+        else m.rollbackTo()
 
-        return true
+        return ok
     }
 
 }
